Move MenuControl base class to TypeScript
